Angular @defer - Section Introduction
After following along with the @defer section, I noticed none of the examples currenly work as they did during the filming of these videos. The @defer blocks of code are not packaged in separate bundles and are not lazy loaded on demand. The defer code in these examples are currently packaged with the main bundles and eager loaded.

In order for the bundling and lazy loading in these examples to work, you must convert your application to a stand-alone component based application. If you jump ahead and do the stand-alone component lessons and then come back to these lessons, they will work properly.

Here is an explanation from the angular documentation:

Which dependencies are defer-loadable?
In order for dependencies within a @defer block to be deferred, they need to meet two conditions:

They must be standalone. Non-standalone dependencies cannot be deferred and will still be eagerly loaded, even inside of @defer blocks.

They must not be directly referenced from the same file, outside of @defer blocks; this includes ViewChild queries.

Transitive dependencies of the components, directives, and pipes used in the defer block can be standalone or NgModule based and will still be deferred.

https://angular.io/guide/defer

*****These lessons (@defer) should be moved after the stand-alone component lessons to avoid this issue***